# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FREGE AI – FregeOS Agricultural Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /farmconnect/posts:
    post:
      operationId: createCommunityPost
      tags: [farmconnect]
      summary: Create a community post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityPostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPost"
  /farmconnect/posts/{postId}/like:
    post:
      operationId: likeCommunityPost
      tags: [farmconnect]
      summary: Like a community post
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPost"
  /farmconnect/nearby-farmers:
    get:
      operationId: getNearbyFarmers
      tags: [farmconnect]
      summary: Get verified farmers nearby
      parameters:
        - in: query
          name: radiusKm
          schema:
            type: number
      responses:
        "200":
          description: Nearby farmers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NearbyFarmer"
  /farmconnect/groups:
    get:
      operationId: getFarmerGroups
      tags: [farmconnect]
      summary: List farmer groups
      responses:
        "200":
          description: Groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FarmerGroup"
    post:
      operationId: createFarmerGroup
      tags: [farmconnect]
      summary: Create a farmer group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmerGroupInput"
      responses:
        "201":
          description: Created group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmerGroup"
  /farmconnect/labour-exchange:
    get:
      operationId: getLabourExchangeListings
      tags: [farmconnect]
      summary: Get labour exchange requests
      responses:
        "200":
          description: Labour requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabourExchange"
    post:
      operationId: createLabourExchange
      tags: [farmconnect]
      summary: Post a labour exchange request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabourExchangeInput"
      responses:
        "201":
          description: Created request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabourExchange"
  /farmconnect/equipment-exchange:
    get:
      operationId: getEquipmentExchange
      tags: [farmconnect]
      summary: Get equipment sharing listings
      responses:
        "200":
          description: Equipment listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EquipmentExchange"
    post:
      operationId: createEquipmentExchange
      tags: [farmconnect]
      summary: Post an equipment for sharing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentExchangeInput"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentExchange"
  /farmconnect/seed-exchange:
    get:
      operationId: getSeedExchange
      tags: [farmconnect]
      summary: Get seed and seedling exchange listings
      responses:
        "200":
          description: Seed exchange listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeedExchange"
    post:
      operationId: createSeedExchange
      tags: [farmconnect]
      summary: Post a seed exchange listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeedExchangeInput"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeedExchange"
  /farmconnect/events:
    get:
      operationId: getCommunityEvents
      tags: [farmconnect]
      summary: Get upcoming community events
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunityEvent"
    post:
      operationId: createCommunityEvent
      tags: [farmconnect]
      summary: Create a community event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityEventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityEvent"
...
components:
  schemas:
...
    CommunityPost:
      type: object
      required: [id, authorId, authorName, authorVerified, content, likes, comments, createdAt]
      properties:
        id:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatar:
          type: ["string", "null"]
        authorVerified:
          type: boolean
        authorReputation:
          type: string
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        likes:
          type: integer
        comments:
          type: integer
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
    CommunityPostInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        imageUrl:
          type: string
        tags:
          type: array
          items:
            type: string
    NearbyFarmer:
      type: object
      required: [id, name, farmType, state, lga, neuroScore, communityReputation, verificationStatus, distanceKm]
      properties:
        id:
          type: integer
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        farmType:
          type: string
        state:
          type: string
        lga:
          type: string
        neuroScore:
          type: number
        communityReputation:
          type: string
        verificationStatus:
          type: string
        distanceKm:
          type: number
    FarmerGroup:
      type: object
      required: [id, name, type, category, description, memberCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [location, crop, livestock]
        category:
          type: string
        description:
          type: string
        memberCount:
          type: integer
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    FarmerGroupInput:
      type: object
      required: [name, type, category, description]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [location, crop, livestock]
        category:
          type: string
        description:
          type: string
    LabourExchange:
      type: object
      required: [id, requesterId, requesterName, title, description, activityType, location, dateNeeded, creditsOffered, status, createdAt]
      properties:
        id:
          type: integer
        requesterId:
          type: integer
        requesterName:
          type: string
        requesterAvatar:
          type: ["string", "null"]
        requesterReputation:
          type: string
        title:
          type: string
        description:
          type: string
        activityType:
          type: string
          enum: [planting, harvesting, land_preparation, irrigation, pest_control, other]
        location:
          type: string
        dateNeeded:
          type: string
        workersNeeded:
          type: integer
        creditsOffered:
          type: integer
        status:
          type: string
          enum: [open, filled, completed, cancelled]
        createdAt:
          type: string
    LabourExchangeInput:
      type: object
      required: [title, description, activityType, location, dateNeeded, workersNeeded, creditsOffered]
      properties:
        title:
          type: string
        description:
          type: string
        activityType:
          type: string
          enum: [planting, harvesting, land_preparation, irrigation, pest_control, other]
        location:
          type: string
        dateNeeded:
          type: string
        workersNeeded:
          type: integer
        creditsOffered:
          type: integer
    EquipmentExchange:
      type: object
      required: [id, ownerId, ownerName, ownerReputation, equipmentType, name, description, pricePerDayNgn, depositNgn, availableFrom, availableTo, status, createdAt]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        ownerName:
          type: string
        ownerAvatar:
          type: ["string", "null"]
        ownerReputation:
          type: string
        equipmentType:
          type: string
          enum: [tractor, harvester, sprayer, water_pump, irrigation, plough, other]
        name:
          type: string
        description:
          type: string
        pricePerDayNgn:
          type: number
        depositNgn:
          type: number
        location:
          type: string
        availableFrom:
          type: string
        availableTo:
          type: string
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, rented, maintenance]
        createdAt:
          type: string
    EquipmentExchangeInput:
      type: object
      required: [equipmentType, name, description, pricePerDayNgn, depositNgn, location, availableFrom, availableTo]
      properties:
        equipmentType:
          type: string
          enum: [tractor, harvester, sprayer, water_pump, irrigation, plough, other]
        name:
          type: string
        description:
          type: string
        pricePerDayNgn:
          type: number
        depositNgn:
          type: number
        location:
          type: string
        availableFrom:
          type: string
        availableTo:
          type: string
        imageUrl:
          type: string
    SeedExchange:
      type: object
      required: [id, sellerId, sellerName, sellerReputation, cropType, variety, listingType, quantityKg, priceNgn, location, status, createdAt]
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        sellerName:
          type: string
        sellerAvatar:
          type: ["string", "null"]
        sellerReputation:
          type: string
        cropType:
          type: string
        variety:
          type: ["string", "null"]
        listingType:
          type: string
          enum: [sell, buy, exchange, donate]
        quantityKg:
          type: number
        priceNgn:
          type: ["number", "null"]
        location:
          type: string
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, completed, expired]
        createdAt:
          type: string
    SeedExchangeInput:
      type: object
      required: [cropType, listingType, quantityKg, location]
      properties:
        cropType:
          type: string
        variety:
          type: string
        listingType:
          type: string
          enum: [sell, buy, exchange, donate]
        quantityKg:
          type: number
        priceNgn:
          type: number
        location:
          type: string
        imageUrl:
          type: string
    CommunityEvent:
      type: object
      required: [id, organizerId, organizerName, title, description, eventType, location, eventDate, rsvpCount, createdAt]
      properties:
        id:
          type: integer
        organizerId:
          type: integer
        organizerName:
          type: string
        title:
          type: string
        description:
          type: string
        eventType:
          type: string
          enum: [training, meeting, market_day, farm_visit, other]
        location:
          type: string
        eventDate:
          type: string
        rsvpCount:
          type: integer
        maxAttendees:
          type: ["integer", "null"]
        createdAt:
          type: string
    CommunityEventInput:
      type: object
      required: [title, description, eventType, location, eventDate]
      properties:
        title:
          type: string
        description:
          type: string
        eventType:
          type: string
          enum: [training, meeting, market_day, farm_visit, other]
        location:
          type: string
        eventDate:
          type: string
        maxAttendees:
          type: integer
...