# lib/api-zod/src/generated/types/farmerGroup.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FREGE AI – FregeOS Agricultural Operating System API
 * OpenAPI spec version: 0.1.0
 */
import type { FarmerGroupType } from './farmerGroupType';

export interface FarmerGroup {
  id: number;
  name: string;
  type: FarmerGroupType;
  category: string;
  description: string;
  memberCount: number;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}